MAT-10220 move sharing library logic to dedicated service to avoid exceeding 750 max line limit. Also, promotes single responsibility principle
