+++ src/test/java/gov/cms/madie/cqllibraryservice/services/CqlLibraryServiceTest.java
@@ -16,7 +16,6 @@ import gov.cms.madie.cqllibraryservice.exceptions.*;
 import gov.cms.madie.cqllibraryservice.locks.CqlLibraryLock;
 import gov.cms.madie.cqllibraryservice.repositories.LibrarySetRepository;
 import gov.cms.madie.models.access.AclSpecification;
-import gov.cms.madie.models.access.RoleEnum;
 import gov.cms.madie.models.common.*;
 import gov.cms.madie.models.dto.LibraryUsage;
 import gov.cms.madie.models.dto.UserDetailsDto;
@@ -38,9 +37,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.PageRequest;
 
-import java.time.Clock;
 import java.time.Instant;
-import java.time.ZoneId;
 import java.time.temporal.ChronoUnit;
 import java.util.*;
 
@@ -54,6 +51,7 @@ import ca.uhn.fhir.rest.server.exceptions.InvalidRequestException;
 class CqlLibraryServiceTest {
 
   @Spy @InjectMocks private CqlLibraryService cqlLibraryService;
+  @Mock private LibrarySharingService librarySharingService;
   @Mock private CqlLibraryRepository cqlLibraryRepository;
   @Mock private LibrarySetService librarySetService;
   @Mock private MeasureServiceClient measureServiceClient;
@@ -356,6 +354,45 @@ class CqlLibraryServiceTest {
             ResourceNotFoundException.class,
             () -> cqlLibraryService.findCqlLibraryById(id, USERNAME));
     assertEquals(ex.getMessage(), "Could not find resource CQL Library with id: " + id);
+  }
+
+  @Test
+  public void testGetSharedLibrariesDelegates() {
+    List<String> libraryIds = List.of("lib1", "lib2");
+    Map<String, List<SharedUser>> expected = Map.of("lib1", List.of(), "lib2", List.of());
+    when(librarySharingService.getSharedLibraries(libraryIds, USERNAME)).thenReturn(expected);
+
+    Map<String, List<SharedUser>> result =
+        cqlLibraryService.getSharedLibraries(libraryIds, USERNAME);
+
+    assertEquals(expected, result);
+    verify(librarySharingService, times(1)).getSharedLibraries(libraryIds, USERNAME);
+  }
+
+  @Test
+  public void testShareLibrariesDelegates() {
+    Map<String, List<String>> input = Map.of("lib1", List.of("user1"));
+    Map<String, List<AclSpecification>> expected = Map.of("lib1", List.of());
+    when(librarySharingService.shareLibraries(input, USERNAME, ACCESSTOKEN)).thenReturn(expected);
+
+    Map<String, List<AclSpecification>> result =
+        cqlLibraryService.shareLibraries(input, USERNAME, ACCESSTOKEN);
+
+    assertEquals(expected, result);
+    verify(librarySharingService, times(1)).shareLibraries(input, USERNAME, ACCESSTOKEN);
+  }
+
+  @Test
+  public void testUnshareLibrariesDelegates() {
+    Map<String, List<String>> input = Map.of("lib1", List.of("user1"));
+    Map<String, List<AclSpecification>> expected = Map.of("lib1", List.of());
+    when(librarySharingService.unshareLibraries(input, USERNAME, ACCESSTOKEN)).thenReturn(expected);
+
+    Map<String, List<AclSpecification>> result =
+        cqlLibraryService.unshareLibraries(input, USERNAME, ACCESSTOKEN);
+
+    assertEquals(expected, result);
+    verify(librarySharingService, times(1)).unshareLibraries(input, USERNAME, ACCESSTOKEN);
   }
 
   @Test
@@ -890,366 +927,6 @@ class CqlLibraryServiceTest {
             eq("setId"), anyBoolean(), eq("L1")))
         .thenReturn(0);
     assertFalse(cqlLibraryService.hasAssociatedLibraries(l1));
-  }
-
-  @Test
-  public void testGetSharedLibrariesWithNoLibraryFound() {
-    String libraryId1 = "libraryId1";
-    List<String> libraryIds = List.of(libraryId1);
-
-    when(cqlLibraryRepository.findById(libraryId1)).thenReturn(Optional.empty());
-
-    assertThrows(
-        ResourceNotFoundException.class,
-        () -> cqlLibraryService.getSharedLibraries(libraryIds, USERNAME));
-  }
-
-  @Test
-  public void testGetSharedLibrariesWithNoLibrarySetFound() {
-    CqlLibrary library1 =
-        CqlLibrary.builder().id("libraryId1").librarySetId("librarySetId1").build();
-
-    when(cqlLibraryRepository.findById("libraryId1")).thenReturn(Optional.ofNullable(library1));
-
-    assertThrows(
-        ResourceNotFoundException.class,
-        () -> cqlLibraryService.getSharedLibraries(List.of("libraryId1"), USERNAME));
-  }
-
-  @Test
-  public void testGetSharedLibrariesWithNoLibrarySetAclsFoundForOneLibrary() {
-    AclSpecification aclSpec = new AclSpecification();
-    aclSpec.setUserId("john");
-    aclSpec.setRoles(
-        new HashSet<>() {
-          {
-            add(RoleEnum.SHARED_WITH);
-          }
-        });
-
-    LibrarySet librarySet1 =
-        LibrarySet.builder()
-            .librarySetId("librarySetId1")
-            .owner("testUser")
-            .acls(
-                new ArrayList<>() {
-                  {
-                    add(aclSpec);
-                  }
-                })
-            .build();
-
-    String libraryId1 = "libraryId1";
-    CqlLibrary library1 =
-        CqlLibrary.builder().id("libraryId1").librarySetId("librarySetId1").build();
-
-    LibrarySet librarySet2 =
-        LibrarySet.builder().librarySetId("librarySetId1").owner("testUser").build();
-
-    String libraryId2 = "libraryId2";
-    CqlLibrary library2 =
-        CqlLibrary.builder()
-            .id(libraryId1)
-            .librarySetId(librarySet1.getLibrarySetId())
-            .librarySet(librarySet2)
-            .build();
-
-    List<String> libraryIds = List.of(libraryId1, libraryId2);
-
-    when(cqlLibraryRepository.findById("libraryId1")).thenReturn(Optional.ofNullable(library1));
-    when(librarySetService.findByLibrarySetId("librarySetId1")).thenReturn(librarySet1);
-    when(cqlLibraryRepository.findById("libraryId2")).thenReturn(Optional.ofNullable(library2));
-    when(userServiceClient.getBulkUserDetails(anyList()))
-        .thenReturn(
-            Map.of(
-                aclSpec.getUserId(),
-                UserDetailsDto.builder().firstName("John").lastName("Doe").build()));
-    when(librarySetService.formatDisplayName(any(), eq(aclSpec.getUserId())))
-        .thenReturn("John Doe (" + aclSpec.getUserId() + ")");
-
-    Map<String, List<SharedUser>> sharedLibraries =
-        cqlLibraryService.getSharedLibraries(libraryIds, USERNAME);
-
-    assertThat(sharedLibraries.size(), is(equalTo(2)));
-
-    assertTrue(sharedLibraries.containsKey(libraryId1));
-    assertThat(sharedLibraries.get(libraryId1).size(), is(equalTo(1)));
-    assertThat(
-        sharedLibraries.get(libraryId1).get(0).getUserId(),
-        is(equalTo(library1.getLibrarySet().getAcls().get(0).getUserId())));
-    assertThat(
-        sharedLibraries.get(libraryId1).get(0).getDisplayName(),
-        is(equalTo("John Doe (" + aclSpec.getUserId() + ")")));
-
-    assertTrue(sharedLibraries.containsKey(libraryId2));
-    assertThat(sharedLibraries.get(libraryId2).size(), is(equalTo(1)));
-  }
-
-  @Test
-  public void testGetSharedLibraries() {
-    AclSpecification aclSpec1 = new AclSpecification();
-    aclSpec1.setUserId("userId1");
-    aclSpec1.setRoles(
-        new HashSet<>() {
-          {
-            add(RoleEnum.SHARED_WITH);
-          }
-        });
-
-    AclSpecification aclSpec2 = new AclSpecification();
-    aclSpec2.setUserId("userId2");
-    aclSpec2.setRoles(
-        new HashSet<>() {
-          {
-            add(RoleEnum.SHARED_WITH);
-          }
-        });
-
-    LibrarySet librarySet2 =
-        LibrarySet.builder()
-            .librarySetId("librarySetId1")
-            .owner("testUser")
-            .acls(
-                new ArrayList<>() {
-                  {
-                    add(aclSpec1);
-                  }
-                })
-            .build();
-
-    LibrarySet librarySet1 =
-        LibrarySet.builder()
-            .librarySetId("librarySetId1")
-            .owner("testUser")
-            .acls(List.of(aclSpec2, aclSpec1))
-            .build();
-
-    String libraryId1 = "libraryId1";
-    CqlLibrary library1 =
-        CqlLibrary.builder()
-            .id(libraryId1)
-            .librarySetId(librarySet1.getLibrarySetId())
-            .librarySet(librarySet1)
-            .build();
-
-    String libraryId2 = "libraryId2";
-    CqlLibrary library2 =
-        CqlLibrary.builder()
-            .id(libraryId1)
-            .librarySetId(librarySet1.getLibrarySetId())
-            .librarySet(librarySet2)
-            .build();
-
-    Instant fixedInstant = Instant.parse("2025-03-17T10:00:00Z");
-    ZoneId utc = ZoneId.of("UTC");
-    Clock fixedClock = Clock.fixed(fixedInstant, utc);
-
-    LibrarySetActionLog librarySetActionLog =
-        LibrarySetActionLog.builder()
-            .actions(
-                List.of(
-                    AccessControlAction.builder()
-                        .sharedWith(aclSpec1.getUserId())
-                        .actionType(ActionType.SHARED)
-                        .performedAt(fixedClock.instant())
-                        .performedBy("performedByUserId")
-                        .build()))
-            .build();
-
-    List<String> libraryIds = List.of(libraryId1, libraryId2);
-
-    when(cqlLibraryRepository.findById("libraryId1")).thenReturn(Optional.ofNullable(library1));
-    when(librarySetService.findByLibrarySetId("librarySetId1")).thenReturn(librarySet1);
-    when(cqlLibraryRepository.findById("libraryId2")).thenReturn(Optional.ofNullable(library2));
-    when(actionLogService.findLibrarySetActionLogByTargetId(anyString()))
-        .thenReturn(librarySetActionLog);
-    when(userServiceClient.getBulkUserDetails(anyList()))
-        .thenReturn(
-            Map.of(
-                aclSpec1.getUserId(),
-                UserDetailsDto.builder().firstName("John").lastName("Doe").build(),
-                aclSpec2.getUserId(),
-                UserDetailsDto.builder().firstName("Jane").lastName("Doe").build()));
-    when(librarySetService.formatDisplayName(any(), eq(aclSpec1.getUserId())))
-        .thenReturn("John Doe (" + aclSpec1.getUserId() + ")");
-    when(librarySetService.formatDisplayName(any(), eq(aclSpec2.getUserId())))
-        .thenReturn("Jane Doe (" + aclSpec2.getUserId() + ")");
-
-    Map<String, List<SharedUser>> sharedLibraries =
-        cqlLibraryService.getSharedLibraries(libraryIds, USERNAME);
-
-    assertThat(sharedLibraries.size(), is(equalTo(2)));
-
-    assertTrue(sharedLibraries.containsKey(libraryId1));
-    assertThat(sharedLibraries.get(libraryId1).size(), is(equalTo(2)));
-    assertThat(
-        sharedLibraries.get(libraryId1).get(0).getUserId(),
-        is(equalTo(library1.getLibrarySet().getAcls().get(0).getUserId())));
-    assertThat(sharedLibraries.get(libraryId1).get(0).getPerformedAt(), is(equalTo(null)));
-    assertThat(
-        sharedLibraries.get(libraryId1).get(0).getDisplayName(),
-        is(equalTo("Jane Doe (" + aclSpec2.getUserId() + ")")));
-    assertThat(sharedLibraries.get(libraryId1).get(1).getUserId(), is(equalTo("userId1")));
-    assertThat(
-        sharedLibraries.get(libraryId1).get(1).getDisplayName(),
-        is(equalTo("John Doe (" + aclSpec1.getUserId() + ")")));
-
-    assertTrue(sharedLibraries.containsKey(libraryId2));
-    assertThat(sharedLibraries.get(libraryId1).size(), is(equalTo(2)));
-    assertThat(
-        sharedLibraries.get(libraryId2).get(0).getUserId(),
-        is(equalTo(library2.getLibrarySet().getAcls().get(0).getUserId())));
-    assertThat(
-        sharedLibraries.get(libraryId2).get(0).getDisplayName(),
-        is(equalTo("Jane Doe (" + aclSpec2.getUserId() + ")")));
-  }
-
-  @Test
-  public void testUpdateSharedLibrariesWithNoLibraryFound() {
-    Map<String, List<String>> libraries = new HashMap<>();
-
-    String libraryId1 = "libraryId1";
-    String libraryId2 = "libraryId2";
-
-    libraries.put(libraryId1, List.of("userId1", "userId2"));
-    libraries.put(libraryId2, List.of("userId2"));
-
-    when(cqlLibraryRepository.findById(eq(libraryId1))).thenReturn(Optional.empty());
-
-    assertThrows(
-        ResourceNotFoundException.class,
-        () -> cqlLibraryService.shareLibraries(libraries, "userName", ACCESSTOKEN));
-  }
-
-  @Test
-  public void testUpdateSharedLibraries() {
-    Map<String, List<String>> libraries = new HashMap<>();
-
-    AclSpecification aclSpec1 = new AclSpecification();
-    aclSpec1.setUserId("testUser");
-    aclSpec1.setRoles(
-        new HashSet<>() {
-          {
-            add(RoleEnum.SHARED_WITH);
-          }
-        });
-
-    AclSpecification aclSpec2 = new AclSpecification();
-    aclSpec2.setUserId("userId2");
-    aclSpec2.setRoles(
-        new HashSet<>() {
-          {
-            add(RoleEnum.SHARED_WITH);
-          }
-        });
-
-    LibrarySet librarySet1 =
-        LibrarySet.builder()
-            .librarySetId("librarySetId1")
-            .owner("testUser")
-            .acls(List.of(aclSpec2, aclSpec1))
-            .build();
-
-    LibrarySet librarySet2 =
-        LibrarySet.builder()
-            .librarySetId("librarySetId1")
-            .owner("testUser")
-            .acls(List.of(aclSpec2, aclSpec1))
-            .build();
-
-    String libraryId1 = "libraryId1";
-
-    CqlLibrary library1 =
-        CqlLibrary.builder()
-            .id(libraryId1)
-            .librarySetId(librarySet1.getLibrarySetId())
-            .librarySet(librarySet1)
-            .build();
-
-    String libraryId2 = "libraryId2";
-    CqlLibrary library2 =
-        CqlLibrary.builder()
-            .id(libraryId1)
-            .librarySetId(librarySet1.getLibrarySetId())
-            .librarySet(librarySet2)
-            .build();
-
-    libraries.put(libraryId1, List.of("testUser", "userId2"));
-    libraries.put(libraryId2, List.of("userId2"));
-
-    when(cqlLibraryRepository.findById("libraryId1")).thenReturn(Optional.ofNullable(library1));
-    when(librarySetService.findByLibrarySetId("librarySetId1")).thenReturn(librarySet1);
-    when(cqlLibraryRepository.findById("libraryId2")).thenReturn(Optional.ofNullable(library2));
-    when(librarySetService.updateLibrarySetAcls(any(), any(), any(), any(Boolean.class)))
-        .thenReturn(librarySet1);
-
-    AclSpecification aclSpecification1 =
-        AclSpecification.builder().userId("testUser").roles(Set.of(RoleEnum.SHARED_WITH)).build();
-    AclSpecification aclSpecification2 =
-        AclSpecification.builder().userId("userId2").roles(Set.of(RoleEnum.SHARED_WITH)).build();
-
-    Map<String, List<AclSpecification>> updatedSharedLibraries =
-        cqlLibraryService.shareLibraries(libraries, "testUser", ACCESSTOKEN);
-    assertThat(updatedSharedLibraries.size(), is(equalTo(2)));
-
-    assertTrue(updatedSharedLibraries.containsKey(libraryId1));
-    assertTrue(updatedSharedLibraries.containsKey(libraryId2));
-
-    assertThat(
-        updatedSharedLibraries.get(libraryId1),
-        is(equalTo(List.of(aclSpecification2, aclSpecification1))));
-
-    assertThat(
-        updatedSharedLibraries.get(libraryId2),
-        is(equalTo(List.of(aclSpecification2, aclSpecification1))));
-  }
-
-  @Test
-  public void testUnshareLibraries() {
-    Map<String, List<String>> libraries = new HashMap<>();
-
-    AclSpecification aclSpec1 = new AclSpecification();
-    aclSpec1.setUserId("testUser");
-    aclSpec1.setRoles(
-        new HashSet<>() {
-          {
-            add(RoleEnum.SHARED_WITH);
-          }
-        });
-
-    LibrarySet librarySet1 =
-        LibrarySet.builder()
-            .librarySetId("librarySetId1")
-            .owner("testUser")
-            .acls(List.of(aclSpec1))
-            .build();
-
-    String libraryId1 = "libraryId1";
-
-    CqlLibrary library1 =
-        CqlLibrary.builder()
-            .id(libraryId1)
-            .librarySetId(librarySet1.getLibrarySetId())
-            .librarySet(librarySet1)
-            .build();
-
-    libraries.put(libraryId1, List.of("testUser", "userId2"));
-
-    when(cqlLibraryRepository.findById("libraryId1")).thenReturn(Optional.ofNullable(library1));
-    when(librarySetService.findByLibrarySetId("librarySetId1")).thenReturn(librarySet1);
-
-    when(librarySetService.updateLibrarySetAcls(any(), any(), any(), any(Boolean.class)))
-        .thenReturn(librarySet1);
-
-    AclSpecification aclSpecification1 =
-        AclSpecification.builder().userId("testUser").roles(Set.of(RoleEnum.SHARED_WITH)).build();
-
-    when(cqlLibraryAccessControlService.hasAdminRole(anyString(), anyString())).thenReturn(true);
-
-    Map<String, List<AclSpecification>> updatedSharedLibraries =
-        cqlLibraryService.unshareLibraries(libraries, "testUser", ACCESSTOKEN);
-    assertThat(updatedSharedLibraries.size(), is(equalTo(1)));
-    assertThat(updatedSharedLibraries.get(libraryId1), is(equalTo(List.of(aclSpecification1))));
   }
 
   @Test
@@ -2025,102 +1702,6 @@ class CqlLibraryServiceTest {
   }
 
   @Test
-  public void testShareLibrariesByAdminUser() {
-    Map<String, List<String>> libraries = new HashMap<>();
-
-    AclSpecification aclSpec1 = new AclSpecification();
-    aclSpec1.setUserId("testUser");
-    aclSpec1.setRoles(
-        new HashSet<>() {
-          {
-            add(RoleEnum.SHARED_WITH);
-          }
-        });
-
-    LibrarySet librarySet1 =
-        LibrarySet.builder()
-            .librarySetId("librarySetId1")
-            .owner("testUser")
-            .acls(List.of(aclSpec1))
-            .build();
-
-    String libraryId1 = "libraryId1";
-
-    CqlLibrary library1 =
-        CqlLibrary.builder()
-            .id(libraryId1)
-            .librarySetId(librarySet1.getLibrarySetId())
-            .librarySet(librarySet1)
-            .build();
-
-    libraries.put(libraryId1, List.of("testUser", "userId2"));
-
-    when(cqlLibraryRepository.findById("libraryId1")).thenReturn(Optional.ofNullable(library1));
-    when(librarySetService.findByLibrarySetId("librarySetId1")).thenReturn(librarySet1);
-
-    when(librarySetService.updateLibrarySetAcls(any(), any(), any(), any(Boolean.class)))
-        .thenReturn(librarySet1);
-
-    AclSpecification aclSpecification1 =
-        AclSpecification.builder().userId("testUser").roles(Set.of(RoleEnum.SHARED_WITH)).build();
-
-    when(cqlLibraryAccessControlService.hasAdminRole(anyString(), anyString())).thenReturn(false);
-
-    Map<String, List<AclSpecification>> updatedSharedLibraries =
-        cqlLibraryService.shareLibraries(libraries, "testUser", ACCESSTOKEN);
-    assertThat(updatedSharedLibraries.size(), is(equalTo(1)));
-    assertThat(updatedSharedLibraries.get(libraryId1), is(equalTo(List.of(aclSpecification1))));
-  }
-
-  @Test
-  public void testUnshareLibrariesByAdminUser() {
-    Map<String, List<String>> libraries = new HashMap<>();
-
-    AclSpecification aclSpec1 = new AclSpecification();
-    aclSpec1.setUserId("testUser");
-    aclSpec1.setRoles(
-        new HashSet<>() {
-          {
-            add(RoleEnum.SHARED_WITH);
-          }
-        });
-
-    LibrarySet librarySet1 =
-        LibrarySet.builder()
-            .librarySetId("librarySetId1")
-            .owner("testUser")
-            .acls(List.of(aclSpec1))
-            .build();
-
-    String libraryId1 = "libraryId1";
-
-    CqlLibrary library1 =
-        CqlLibrary.builder()
-            .id(libraryId1)
-            .librarySetId(librarySet1.getLibrarySetId())
-            .librarySet(librarySet1)
-            .build();
-
-    libraries.put(libraryId1, List.of("testUser", "userId2"));
-
-    when(cqlLibraryRepository.findById("libraryId1")).thenReturn(Optional.ofNullable(library1));
-    when(librarySetService.findByLibrarySetId("librarySetId1")).thenReturn(librarySet1);
-
-    when(librarySetService.updateLibrarySetAcls(any(), any(), any(), any(Boolean.class)))
-        .thenReturn(librarySet1);
-
-    AclSpecification aclSpecification1 =
-        AclSpecification.builder().userId("testUser").roles(Set.of(RoleEnum.SHARED_WITH)).build();
-
-    when(cqlLibraryAccessControlService.hasAdminRole(anyString(), anyString())).thenReturn(true);
-
-    Map<String, List<AclSpecification>> updatedSharedLibraries =
-        cqlLibraryService.unshareLibraries(libraries, "testUser", ACCESSTOKEN);
-    assertThat(updatedSharedLibraries.size(), is(equalTo(1)));
-    assertThat(updatedSharedLibraries.get(libraryId1), is(equalTo(List.of(aclSpecification1))));
-  }
-
-  @Test
   public void testUpdateCqlLibraryWhenUserNameIsNull() {
     CqlLibrary library = CqlLibrary.builder().id("libraryId").build();
 
@@ -2164,26 +1745,6 @@ class CqlLibraryServiceTest {
                 true,
                 "Info",
                 "test-okta"));
-  }
-
-  @Test
-  public void testGetSharedLibrariesLibrarySetAclsNull() {
-    String libraryId = "libraryId";
-    LibrarySet librarySet = LibrarySet.builder().librarySetId("librarySetId").build();
-    CqlLibrary library =
-        CqlLibrary.builder()
-            .id(libraryId)
-            .librarySetId("librarySetId")
-            .librarySet(librarySet)
-            .build();
-    when(cqlLibraryRepository.findById(libraryId)).thenReturn(Optional.of(library));
-    when(librarySetService.findByLibrarySetId("librarySetId")).thenReturn(librarySet);
-
-    Map<String, List<SharedUser>> result =
-        cqlLibraryService.getSharedLibraries(List.of(libraryId), "test-okta");
-    assertNotNull(result);
-    assertTrue(result.containsKey(libraryId));
-    assertTrue(result.get(libraryId).isEmpty());
   }
 
   @Test
